some changes in lab_2_2 and lab_2_3 (ELEVATORS ARE WORKING!!!)

+++ semester 2/lab_2_2/libs/bigfloat.h
@@ -24,7 +24,7 @@ public:
 	bigfloat(bigint value);
 
 	bigfloat(bigint numerator, bigint denominator);
-	bigfloat(double numerator, double denominator);
+	bigfloat(int numerator, int denominator);
 
 	bigfloat operator-() const;
 	bigfloat& negate();
